requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
djoser==2.2.2

psycopg2-binary==2.9.9
environs==10.3.0
drf-yasg==1.21.7
gunicorn==21.2.0

pillow==10.2.0

django-silk==5.1.0
django-debug-toolbar==4.3.0

#black==23.12.1
#flake8==6.1.0
#isort==5.13.2
#pre-commit==3.5.0